+++ Src/Web/Web.Shared.Net/ApplicationInsightsHttpModule.cs
@@ -103,9 +103,6 @@
 
                     this.requestModule.OnBeginRequest(httpApplication.Context);
                 }
-
-                // Kept for backcompat. Should be removed in 2.3 SDK
-                WebEventsPublisher.Log.OnBegin();
             }
         }
 
@@ -183,10 +180,6 @@
                     {
                         this.requestModule.OnEndRequest(httpApplication.Context);
                     }
-
-                    // Kept for backcompat. Should be removed in 2.3 SDK
-                    WebEventsPublisher.Log.OnError();
-                    WebEventsPublisher.Log.OnEnd();
                 }
                 else
                 {