+++ CHANGELOG.md
@@ -6,9 +6,13 @@
 - Implemented [Correlation HTTP protocol](https://github.com/lmolkova/correlation/blob/master/http_protocol_proposal_v1.md): default headers to pass Operation Root Id and Parent Id were changed. This version is backward compatible with previously supported headers. 
 - Improvements to exception statistics, e.g. 2 of each type of exception will be output via TrackException
 
+## Version 2.3.0
+- Includes all changes since 2.2.0 stable release.
+- Exception statistics feature introduced in beta version is removed.
 
 ## Version 2.3.0-beta3
 - Exception statistics improvements and other minor bug fixes. [Full list.] (https://github.com/Microsoft/ApplicationInsights-dotnet-server/milestone/19?closed=1)
+- Cross Components Correlation ID changed from SHA(instrumentation key) to Application ID retrieved from http endpoint `api/profiles/{ikey}/appId`.
 
 ## Version 2.3.0-beta2
 - Automatic collection of first chance exceptions statistics. Use a query like this in Application Analytics to query for this statistics: